Refactor useToastNotification async toasts
